requirements.txt
pyyaml
pandas
RPi-GPIO
himl